Move version command to separate package
